[TASK] Separate definition error handling in TCA
Changes the way the definition error is handled within entity
notifications.

The old way consisted in adding display conditions to every field to
check if the definition contains error.

Now, a data provider does the same job, and replaces the whole TCA if an
error is found.
